Refactored tmapi test fixed readme lines
