+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "env:test": "NODE_ENV=testing npx playwright test shoppingSite.spec.ts",
     "api:mock": "NODE_ENV=testing npx playwright test shoppingApiMock.spec.ts",
-    "env:tm": "NODE_ENV=test npx playwright test TradeMe.spec.ts tradmeApi.spec.ts  --project=chromium","env:tm2": "NODE_ENV=trademe npx playwright test tradmeApi.spec.ts --project=chromium",
+    "env:tm": "NODE_ENV=trademe npx playwright test TradeMe.spec.ts tradmeApi.spec.ts  --project=chromium",
     "env:dev": "NODE_ENV=development npx playwright test",
     "test:env": "ts-node test-env.ts"
   },
